% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.arg_check_herbarium}
\alias{.arg_check_herbarium}
\title{Validate herbarium codes}
\usage{
.arg_check_herbarium(x)
}
\arguments{
\item{x}{Character vector of herbarium codes.}
}
\value{
Invisibly TRUE.
}
\description{
Validate herbarium codes
}
\keyword{internal}
