% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.best_numeric_vec}
\alias{.best_numeric_vec}
\title{Choose the best numeric candidate column by completeness}
\usage{
.best_numeric_vec(df, candidates, default = NA_real_)
}
\arguments{
\item{df}{Data frame.}

\item{candidates}{Character vector or list of character vectors.}

\item{default}{Default numeric value.}
}
\value{
Vector of length `nrow(df)`.
}
\description{
Choose the best numeric candidate column by completeness
}
\keyword{internal}
