% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.calculate_phytosociological_metrics}
\alias{.calculate_phytosociological_metrics}
\title{Calculate phytosociological metrics from canonical field-sheet data}
\usage{
.calculate_phytosociological_metrics(
  fp_sheet,
  plot_size_ha = 1,
  subplot_size_m = 10
)
}
\arguments{
\item{fp_sheet}{Canonical field-sheet data frame.}

\item{plot_size_ha}{Plot size in hectares.}

\item{subplot_size_m}{Subplot side in meters.}
}
\value{
Named list with species, family and diversity metrics.
}
\description{
Calculate phytosociological metrics from canonical field-sheet data
}
\keyword{internal}
