% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.clean_chr}
\alias{.clean_chr}
\title{Return a cleaned text vector}
\usage{
.clean_chr(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Character vector.
}
\description{
Return a cleaned text vector
}
\keyword{internal}
