% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.duckdb_load_occurrence}
\alias{.duckdb_load_occurrence}
\title{Load occurrence.txt into DuckDB}
\usage{
.duckdb_load_occurrence(
  con,
  occurrence_path,
  herbarium,
  source,
  resource_id,
  verbose = FALSE
)
}
\arguments{
\item{con}{DuckDB connection.}

\item{occurrence_path}{Character path to occurrence.txt.}

\item{herbarium}{Herbarium code.}

\item{source}{One of "jabot" or "reflora".}

\item{resource_id}{IPT resource id.}

\item{verbose}{Logical.}
}
\value{
Logical.
}
\description{
Load occurrence.txt into DuckDB
}
\keyword{internal}
