% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.extract_number_token}
\alias{.extract_number_token}
\title{Extract a collector number token from free text}
\usage{
.extract_number_token(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with extracted number token.
}
\description{
Extract a collector number token from free text
}
\keyword{internal}
