% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.field_sheet_cols}
\alias{.field_sheet_cols}
\title{Standard field-sheet schema used internally by plot functions}
\usage{
.field_sheet_cols()
}
\value{
Character vector of canonical column names.
}
\description{
Standard field-sheet schema used internally by plot functions
}
\keyword{internal}
