% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.get_latest_version_info}
\alias{.get_latest_version_info}
\title{Read latest version info from an IPT resource page}
\usage{
.get_latest_version_info(resource_url)
}
\arguments{
\item{resource_url}{Character scalar.}
}
\value{
Named list with version, published_on and records.
}
\description{
Read latest version info from an IPT resource page
}
\keyword{internal}
