% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.has_any}
\alias{.has_any}
\title{Test whether any alias exists in a data frame}
\usage{
.has_any(df, candidates)
}
\arguments{
\item{df}{Data frame.}

\item{candidates}{Character vector.}
}
\value{
Logical scalar.
}
\description{
Test whether any alias exists in a data frame
}
\keyword{internal}
