% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.herbaria_db_connect}
\alias{.herbaria_db_connect}
\title{Connect to herbarium DuckDB cache}
\usage{
.herbaria_db_connect()
}
\value{
DuckDB connection.
}
\description{
Connect to herbarium DuckDB cache
}
\keyword{internal}
