% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.make_jabot_url}
\alias{.make_jabot_url}
\title{Build a JABOT specimen URL}
\usage{
.make_jabot_url(catalogNumber, herbarium)
}
\arguments{
\item{catalogNumber}{Character scalar.}

\item{herbarium}{Character scalar.}
}
\value{
Character scalar URL or NA.
}
\description{
Build a JABOT specimen URL
}
\keyword{internal}
