% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.make_reflora_url}
\alias{.make_reflora_url}
\title{Build a REFLORA image URL}
\usage{
.make_reflora_url(occurrenceID)
}
\arguments{
\item{occurrenceID}{Character scalar. Aceita URI completa ou ID numerico puro.}
}
\value{
Character scalar URL or NA.
}
\description{
Build a REFLORA image URL
}
\keyword{internal}
