% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.norm_nm}
\alias{.norm_nm}
\title{Normalize names for matching}
\usage{
.norm_nm(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector normalized to lowercase ASCII alphanumeric text.
}
\description{
Normalize names for matching
}
\keyword{internal}
