% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.normalize_number_token}
\alias{.normalize_number_token}
\title{Keep only digits in collector numbers}
\usage{
.normalize_number_token(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with only digits.
}
\description{
Keep only digits in collector numbers
}
\keyword{internal}
