% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.normalize_station_name}
\alias{.normalize_station_name}
\title{Normalize station names}
\usage{
.normalize_station_name(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Character vector.
}
\description{
Normalize station names
}
\keyword{internal}
