% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.normalize_station_number}
\alias{.normalize_station_number}
\title{Normalize station numeric identifiers}
\usage{
.normalize_station_number(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Character vector.
}
\description{
Normalize station numeric identifiers
}
\keyword{internal}
