% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.parse_num}
\alias{.parse_num}
\title{Parse numeric text robustly}
\usage{
.parse_num(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Numeric vector.
}
\description{
Parse numeric text robustly
}
\keyword{internal}
