% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.parse_year}
\alias{.parse_year}
\title{Parse year values from mixed text}
\usage{
.parse_year(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Integer vector.
}
\description{
Parse year values from mixed text
}
\keyword{internal}
