% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.prepare_report_dashboard}
\alias{.prepare_report_dashboard}
\title{Prepare taxonomic dashboard tables and plots}
\usage{
.prepare_report_dashboard(
  fp_sheet,
  plot_size_ha = 1,
  subplot_size_m = 10,
  language = "en"
)
}
\arguments{
\item{fp_sheet}{Canonical field-sheet data frame.}

\item{plot_size_ha}{Plot size in hectares.}

\item{subplot_size_m}{Subplot size in meters.}

\item{language}{Language code.}
}
\value{
Named list with summary tables and plots.
}
\description{
Prepare taxonomic dashboard tables and plots
}
\keyword{internal}
