% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.read_best_plot_sheet}
\alias{.read_best_plot_sheet}
\title{Read the best worksheet from a workbook}
\usage{
.read_best_plot_sheet(path, sheet = NULL)
}
\arguments{
\item{path}{File path.}

\item{sheet}{Optional preferred sheet name or index.}
}
\value{
Data frame with `sheet_name` attribute.
}
\description{
Read the best worksheet from a workbook
}
\keyword{internal}
