% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.resource_guess_herbarium}
\alias{.resource_guess_herbarium}
\title{Convert one IPT resource id into a guessed herbarium code}
\usage{
.resource_guess_herbarium(resource_id)
}
\arguments{
\item{resource_id}{Character scalar.}
}
\value{
Character scalar.
}
\description{
Convert one IPT resource id into a guessed herbarium code
}
\keyword{internal}
