% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.score_plot_sheet}
\alias{.score_plot_sheet}
\title{Score workbook sheets for plot data detection}
\usage{
.score_plot_sheet(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
Integer score.
}
\description{
Score workbook sheets for plot data detection
}
\keyword{internal}
