% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{.split_recordedby_people}
\alias{.split_recordedby_people}
\title{Split recordedBy into candidate people}
\usage{
.split_recordedby_people(recordedBy)
}
\arguments{
\item{recordedBy}{Character scalar.}
}
\value{
Character vector of candidate person strings.
}
\description{
Split recordedBy into candidate people
}
\keyword{internal}
