% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TUM}
\alias{TUM}
\title{Countries and associated classification of botanical divisions}
\usage{
data(TUM)
}
\description{
Countries and associated classification of botanical divisions according to
the World Geographical Scheme for Recording Plant Distributions (WGSRPD).
This dataset provides the tabular look-up table with multilingual names for
all botanical regions recognized by the standard.
}
\examples{
\dontrun{
library(forplotR)
data(TUM)
head(TUM)
}

}
