% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_for_balance.R
\name{.create_rmd_content}
\alias{.create_rmd_content}
\title{Prepare RMarkdown sections for each subplot with navigation and page breaks}
\usage{
.create_rmd_content(
  subplot_plots,
  tf_col,
  tf_uncol,
  tf_palm,
  plot_name,
  plot_code
)
}
\description{
Prepare RMarkdown sections for each subplot with navigation and page breaks
}
