% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_for_balance.R
\name{plot_for_balance}
\alias{plot_for_balance}
\title{Generate forest plot specimen map and collection balance}
\usage{
plot_for_balance(fp_file_path = NULL,
                 plot_size = 1,
                 subplot_size = 10,
                 highlight_palms = TRUE,
                 dir = "Results_map_plot",
                 filename = "plot_specimen")
}
\arguments{
\item{fp_file_path}{File path to the forestplots dataset (Excel format).}

\item{plot_size}{Overall plot size in hectares (default is 1 ha).}

\item{subplot_size}{Side length of subplots in meters (default is 10 m).}

\item{highlight_palms}{Logical. If TRUE, highlights Arecaceae in purple.}

\item{dir}{Directory path where output will be saved
(default is "Results_map_plot").}

\item{filename}{Name of the output file for the main plot (without extension).}
}
\value{
- A full PDF report with:
  - plot metadata,
  - navigable sections including subplot maps;
- An Excel spreadsheet with:
  - collection percentages (overall and per subplot),
  - tag numbers for collected and not collected individuals.
}
\description{
Processes field data collected using the \href {https://forestplots.net/}{Forestplots format}
and generates a specimen map (2D plot) with collection status and spatial
distribution of individuals across subplots. Generate a PDF with a full plot
report with separate maps, for each subplot and a spreadsheet summarizing the
percentage of collected specimens per subplot. The function performs the
following steps: (i) validates all input arguments and checks/create output
folders; (ii) reads the xlsx {Forestplots format} sheet, extracts metadata
(team, plot name, plot code), and cleans the data; (iii) normalizes and cleans
coordinate and diameter values, computing global plot coordinates from
subplot-relative positions; (iv) creates a PDF report, including plot metadata,
the main map, collected and uncollected specimen maps, optionally a map
highlighting palms (Arecaceae) and navigable individual subplot  maps; (v)
optionally generates a xlsx spreadsheet with the collection
percentage per subplot (including totals), distinguishing collected vs.
uncollected, and palms specimen.
}
\examples{
\dontrun{
plot_for_balance(fp_file_path = "data/forestplot.xlsx",
                 plot_size = 1,
                 subplot_size = 10,
                 highlight_palms = TRUE,
                 dir = "Results_map_plot",
                 filename = "plot_specimen")
}

}
\author{
Giulia Ottino & Domingos Cardoso
}
