% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_html_map.R
\name{plot_html_map}
\alias{plot_html_map}
\title{Plot and save vouchered forest plot map as HTML}
\usage{
plot_html_map(fp_file_path = NULL,
                     vertex_coords = NULL,
                     plot_size = 1,
                     subplot_size = 10,
                     map_type = c("satellite", "street"),
                     voucher_imgs = NULL,
                     filename = "plot_map.html")
}
\arguments{
\item{fp_file_path}{Character. Path to the Excel file containing tree data.}

\item{vertex_coords}{Data frame or path to Excel file with columns `Latitude`
and `Longitude` for the four plot corners.}

\item{plot_size}{Numeric. Total plot size in hectares (default is 1).}

\item{subplot_size}{Numeric. Subplot size in meters (default is 10).}

\item{map_type}{Character. Base map type: "satellite" or "street" (default
is "satellite").}

\item{voucher_imgs}{Character. Directory path where voucher images are stored.}

\item{filename}{Character. Name of the output HTML file (default is
"plot_map.html").}
}
\description{
Generates an interactive and self-contained HTML map following the
\href{https://forestplots.net/}{ForestPlots plot protocol}, using Leaflet
and based on vouchered tree data and the geographical coordinates of the
plot corners. The function:(i) Converts local subplot coordinates (X, Y) to
geographic coordinates using geospatial interpolation from four corner vertices;
(ii) automatically extracts metadata like team name, plot name, and plot code
from the input xlsx file; (iii) draws the plot boundary polygon with angular
ordering to prevent self-intersection, over a selectable basemap ("satellite"
or "street"); (iv) colors tree points based on collection status and palms:
palms (yellow), collected (gray), and missing (red); (v) embeds image
carousels in the popups for vouchers when corresponding images are available
in the specified folder; (vi) adds an interactive filter sidebar with
checkboxes, multi-select inputs, and search box for filtering by family,
species, collection status, and presence of photos; (vii) displays informative
popups for each tree,showing tag number, subplot,family, species, DBH,
voucher code, and photos (if present); (viii) saves the resulting map
as a date-stamped standalone HTML file to the specified directory for easy
sharing, archiving, or field consultation.
}
\examples{
\dontrun{
vertex_df <- data.frame(
  Latitude = c(-3.123, -3.123, -3.125, -3.125),
  Longitude = c(-60.012, -60.010, -60.012, -60.010)
)
plot_html_map("data/tree_data.xlsx",
               vertex_coords = vertex_df,
               voucher_imgs = "voucher_imgs")
}

}
\author{
Giulia Ottino & Domingos Cardoso
}
